allow semi-reserved keywords in class member names
Make it possible to use semi-reserved keywords for class
methods and constants.

A new GenTree::get_identifier() method is used in all places
where we allow such keywords usage.

Some code in lexer hack_last_tokens() method is re-arranged,
so elseif token is not replaced with else+if tokens in
places where we want it to stay a single elseif token.

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>